Merge/port YAML showcases
