+++ docs/content/format-crate-matrix.md
@@ -16,30 +16,18 @@ Note: `msgp` = `facet-msgpack` (shortened for column width)
 
 Note: S = Serialization, D = Deserialization
 
-| Crate | Direction | Format Type | Parser | Showcase | Error Showcase |
-|-------|-----------|-------------|--------|----------|----------------|
-| [facet-json](https://docs.rs/facet-json) | SD | Text | Event-based (custom) | ✅ | ✅ |
-| [facet-kdl](https://docs.rs/facet-kdl) | SD | Text (node-based) | DOM ([kdl-rs](https://docs.rs/kdl)) | ✅ | ✅ |
-| [facet-yaml](https://docs.rs/facet-yaml) | SD | Text | Event-based ([saphyr](https://docs.rs/saphyr)) | ✅ | ✅ |
-| [facet-toml](https://docs.rs/facet-toml) | SD | Text | DOM ([toml_edit](https://docs.rs/toml_edit)) | 🚫 | 🚫 |
-| [facet-msgpack](https://docs.rs/facet-msgpack) | SD | Binary | Event-based (custom) | 🚫 | 🚫 |
-| [facet-asn1](https://docs.rs/facet-asn1) | S | Binary | (Custom) | 🚫 | 🚫 |
-| [facet-xdr](https://docs.rs/facet-xdr) | S | Binary | (Custom) | 🚫 | 🚫 |
-| [facet-args](https://docs.rs/facet-args) | D | CLI | (Custom) | 🚫 | 🚫 |
-| [facet-urlencoded](https://docs.rs/facet-urlencoded) | D | Text | (Custom) | 🚫 | 🚫 |
-| [facet-csv](https://docs.rs/facet-csv) | S | Text | (Custom) | 🚫 | 🚫 |
-
-## API Surface
-
-| Feature | json | kdl | yaml | toml | msgp | asn1 | xdr | args | urlenc | csv |
-|---------|------|-----|------|------|------|------|-----|------|--------|-----|
-| `from_str` | ✅ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ | ✅ | ✅ | ➖ |
-| `from_slice` | ➖ | ➖ | ➖ | ➖ | ✅ | ➖ | ➖ | ✅ | ➖ | ➖ |
-| `to_string` | ✅ | ✅ | ✅ | ✅ | ➖ | 🚫 | 🚫 | ➖ | ➖ | ✅ |
-| `to_vec` | ➖ | ➖ | ➖ | ➖ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ |
-| `to_writer` | ✅ | ✅ | ✅ | 🚫 | ✅ | 🚫 | 🚫 | ➖ | ➖ | ✅ |
-| `to_string_pretty` | ✅ | 🚫 | 🚫 | 🚫 | ➖ | ➖ | ➖ | ➖ | ➖ | ➖ |
-| `miette::Diagnostic` | ✅ | ✅ | ✅ | 🟡 | 🚫 | 🚫 | 🚫 | ✅ | 🚫 | 🚫 |
+| Crate | Direction | Format Type | Parser | Showcase |
+|-------|-----------|-------------|--------|----------|
+| [facet-json](https://docs.rs/facet-json) | SD | Text | Event-based (custom) | [View](/showcases/json) |
+| [facet-kdl](https://docs.rs/facet-kdl) | SD | Text (node-based) | DOM ([kdl-rs](https://docs.rs/kdl)) | [View](/showcases/kdl) |
+| [facet-yaml](https://docs.rs/facet-yaml) | SD | Text | Event-based ([saphyr](https://docs.rs/saphyr)) | [View](/showcases/yaml) |
+| [facet-toml](https://docs.rs/facet-toml) | SD | Text | DOM ([toml_edit](https://docs.rs/toml_edit)) | 🚫 |
+| [facet-msgpack](https://docs.rs/facet-msgpack) | SD | Binary | Event-based (custom) | 🚫 |
+| [facet-asn1](https://docs.rs/facet-asn1) | S | Binary | (Custom) | 🚫 |
+| [facet-xdr](https://docs.rs/facet-xdr) | S | Binary | (Custom) | 🚫 |
+| [facet-args](https://docs.rs/facet-args) | D | CLI | (Custom) | 🚫 |
+| [facet-urlencoded](https://docs.rs/facet-urlencoded) | D | Text | (Custom) | 🚫 |
+| [facet-csv](https://docs.rs/facet-csv) | S | Text | (Custom) | 🚫 |
 
 ## Scalar Types
 
@@ -125,15 +113,6 @@ All formats support `String`, `&str` (with best-effort borrowing), and `Cow<str>
 // Adjacently tagged: { "tag": "Variant", "content": "value" }
 ```
 
-### Variant Types
-
-| Type | json | kdl | yaml | toml | msgp | asn1 | xdr | args | urlenc | csv |
-|------|------|-----|------|------|------|------|-----|------|--------|-----|
-| Unit variants | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ |
-| Newtype variants | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ |
-| Tuple variants | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ |
-| Struct variants | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ | ➖ | ➖ | ➖ |
-
 ## Attributes
 
 | Attribute | json | kdl | yaml | toml | msgp | asn1 | xdr | args | urlenc | csv |
@@ -155,6 +134,7 @@ All formats support `String`, `&str` (with best-effort borrowing), and `Cow<str>
 
 | Feature | json | kdl | yaml | toml | msgp | asn1 | xdr | args | urlenc | csv |
 |---------|------|-----|------|------|------|------|-----|------|--------|-----|
+| `miette::Diagnostic` | ✅ | ✅ | ✅ | 🟡 | 🚫 | 🚫 | 🚫 | ✅ | 🚫 | 🚫 |
 | `Spanned<T>` wrapper | ✅ | ✅ | ✅ | 🚫 | 🚫 | 🚫 | 🚫 | ✅ | 🚫 | 🚫 |
 | Solver integration | ✅ | ✅ | 🟡 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 |
 | "Did you mean?" suggestions | ✅ | ✅ | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 | 🚫 |
