+++ docs/templates/page.html
@@ -38,9 +38,8 @@
         {% if page.path is starting_with("/showcases/") %}
         <nav class="format-nav">
             <a href="/showcases/kdl/" {% if page.path == "/showcases/kdl/" %}class="active"{% endif %}>KDL</a>
-            <a href="/showcases/kdl-errors/" {% if page.path == "/showcases/kdl-errors/" %}class="active"{% endif %}>KDL Errors</a>
-            <a href="/showcases/json-errors/" {% if page.path == "/showcases/json-errors/" %}class="active"{% endif %}>JSON Errors</a>
-            <a href="/showcases/yaml-errors/" {% if page.path == "/showcases/yaml-errors/" %}class="active"{% endif %}>YAML Errors</a>
+            <a href="/showcases/json/" {% if page.path == "/showcases/json/" %}class="active"{% endif %}>JSON</a>
+            <a href="/showcases/yaml/" {% if page.path == "/showcases/yaml/" %}class="active"{% endif %}>YAML</a>
         </nav>
         {% endif %}
         <h1 class="page-title">{{ page.title }}</h1>
